+++ OneBlink.SDK/models/FormsAppBase.cs
@@ -80,5 +80,9 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
+        public List<string> groups
+        {
+            get; set;
+        }
     }
 }